Code climate fixes
