+++ src/ui/components/views/gamedetails/body/playtimecard.js
@@ -1,14 +1,11 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import classnames from 'classnames';
 import { withStyles } from 'material-ui/styles';
 import Typography from 'material-ui/Typography';
-import IconButton from 'material-ui/IconButton';
-import ExpandMoreIcon from 'material-ui-icons/ExpandMore';
-import Card, { CardHeader, CardContent, CardActions } from 'material-ui/Card';
+import Card, { CardHeader, CardContent } from 'material-ui/Card';
 import Collapse from 'material-ui/transitions/Collapse';
 import TimesheetList from './timesheetlist';
-
+import ExpandAction from './expandaction';
 
 const AllTimesheets = ({ isOpen }) => (
   <Collapse in={isOpen} timeout="auto" unmountOnExit>
@@ -20,15 +17,7 @@ const AllTimesheets = ({ isOpen }) => (
 AllTimesheets.propTypes = {
   isOpen: PropTypes.bool.isRequired,
 };
-const styles = theme => ({ expand: {
-  transform: 'rotate(0deg)',
-  transition: theme.transitions.create('transform', {
-    duration: theme.transitions.duration.shortest,
-  }),
-},
-  expandOpen: {
-    transform: 'rotate(180deg)',
-  } });
+const styles = () => ({ });
 const propTypes = {
   classes: PropTypes.shape().isRequired,
 };
@@ -55,16 +44,10 @@ class PlaytimeCard extends React.Component {
           You spent a total of 15 hours on this game
         </Typography>
       </CardContent>
-      <CardActions disableActionSpacing>
-        <IconButton
-          className={classnames(classes.expand, { [classes.expandOpen]: isOpen })}
-          onClick={this.handleExpandClick}
-          aria-expanded={isOpen}
-          aria-label="Show more"
-        >
-          <ExpandMoreIcon />
-        </IconButton>
-      </CardActions>
+      <ExpandAction
+        isOpen={isOpen}
+        handleExpandClick={this.handleExpandClick}
+      />
       <AllTimesheets isOpen={isOpen} />
     </Card>);
   }
