+++ src/ui/components/views/gamedetails/body/ratingscard.js
@@ -1,25 +1,14 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import classnames from 'classnames';
 import { withStyles } from 'material-ui/styles';
 import Typography from 'material-ui/Typography';
-import IconButton from 'material-ui/IconButton';
-import ExpandMoreIcon from 'material-ui-icons/ExpandMore';
-import Card, { CardHeader, CardContent, CardActions } from 'material-ui/Card';
+import Card, { CardHeader, CardContent } from 'material-ui/Card';
 import Collapse from 'material-ui/transitions/Collapse';
 import Rating from 'components/widgets/rating';
+import ExpandAction from './expandaction';
 
 
-const styles = theme => ({
-  expand: {
-    transform: 'rotate(0deg)',
-    transition: theme.transitions.create('transform', {
-      duration: theme.transitions.duration.shortest,
-    }),
-  },
-  expandOpen: {
-    transform: 'rotate(180deg)',
-  },
+const styles = () => ({
 });
 const propTypes = {
   classes: PropTypes.shape().isRequired,
@@ -60,16 +49,10 @@ class RatingsCard extends React.Component {
           <span className="ratings-text">4.1/5 <span className="reviews-number">(33 reviews)</span></span>
         </Typography>
       </CardContent>
-      <CardActions disableActionSpacing>
-        <IconButton
-          className={classnames(classes.expand, { [classes.expandOpen]: isOpen })}
-          onClick={this.handleExpandClick}
-          aria-expanded={isOpen}
-          aria-label="Show more"
-        >
-          <ExpandMoreIcon />
-        </IconButton>
-      </CardActions>
+      <ExpandAction
+        isOpen={isOpen}
+        handleExpandClick={this.handleExpandClick}
+      />
       <AllReviews isOpen={isOpen} />
     </Card>);
   }
